requirements.txt
aiohttp==3.10.11
python-dotenv==1.0.0
psycopg2-binary~=2.9.3